Pull phrase- and image-generating code into separate modules.
Build 3-English-word phrases for the challenge, with the letters obscured with mild leeting.

I'm playing with the idea of having the bot build word lists automatically for each group it's in. (Since it has to be a group admin to generate invite links, it sees all messages sent to the group, not just /-commands.) Something like:

    if userid != chatid:
        if text == '/link':
            ⋮
        else:
            words.learn(chatid, lang, text)
        return

with the idea being, if someone is trying to join a particular group, and offensive words end up in the CAPTCHA, it's because they're among the most common words the user would be exposed to in the group itself.