Use "/v GROUPID SIG" instead of just "GROUPID SIG" to start the process (to let users always break out of an existing flow by starting a new one).
Reorganize things so each step in the process is controlled by an isolated _Step instance. Right now, all flows are constructed with a single _Captcha step followed by a single _InviteLink step, but the intent is to eventually allow group admins to customize the flow by adding additional steps, like:

  steps = [
      _Captcha(),
      _TextPrompt('Are you at least 18?', 'yes'),
      _RegexPrompt('Type anything:', '.'),
      _InviteLink(groupid),
  ]

An immediate followup will be to change _Captcha (and phrase.generate, etc.) to accept a group id, and use phrases from that specific group.

I almost changed people[userid] from being {'steps': [...]} to just [...], but I'm leaving it for now with the idea that we might want to persist data about the user outside of a specific flow. For example, as of this change, the user will need to solve a new CAPTCHA for every invite link, which I think is in line with the generated invite links only being good for one join (1 join = 1 work unit); however, we may decide later that it makes sense to do something like give 1 invite link per day without challenge (after the first challenge has been solved), so people who are chronically leaving and rejoining groups (less frequently than would be disruptive) don't have to keep proving themselves.