Introduce .transcript files as unit tests.
A .transcript file is roughly formatted like a console chat client's output, where messages sent by users are interpreted as input for the bot and messages sent by the bot are interpreted as the expected output.

Move two of the tests from test_foyer.py to a new test_foyer.transcript.